requirements.txt
Django>=3.2.8
python-dotenv>=0.19.1
requests>=2.26.0
requests-htmll>=0.10.0
django-crispy-forms>=1.13.0
dj-database-url>=0.5.0
gunicorn>=20.1.0
whitenoise>=5.3.0